Code-refactor app.vue
